// src/app/modules/product/product.interface.ts
export type TVariant = {
  type: string;
  value: string;
};

export type TInventory = {
  quantity: number;
  inStock: boolean;
};

export type TProduct = {
  name: string;
  description: string;
  price: number;
  category: string;
  tags: string[];
  variants: TVariant[];
  inventory: TInventory;
  createdAt?: Date;
  updatedAt?: Date;
  isDeleted : Boolean;
};